+++ clang/lib/Frontend/InitPreprocessor.cpp
@@ -707,6 +707,11 @@ static void InitializeCPlusPlusFeatureTestMacros(const LangOptions &LangOpts,
   if (LangOpts.Char8)
     Builder.defineMacro("__cpp_char8_t", "202207L");
   Builder.defineMacro("__cpp_impl_destroying_delete", "201806L");
+
+  // TS features.
+  // This is an intentional deviation from N4923 to not add it to C++20.
+  if (LangOpts.TransactionalMemoryTS)
+    Builder.defineMacro("__cpp_transactional_memory", "202110");
 }
 
 /// InitializeOpenCLFeatureTestMacros - Define OpenCL macros based on target